+++ cooking/models/saved.py
@@ -46,8 +46,8 @@ event.listen(Saved, 'before_insert', before_insert_listener)
 
 
 saved = Table('saved', metadata,
-    Column('user_id', BIGINT, ForeignKey('users.id', ondelete="CASCADE")),
-    Column('recipe_id', BIGINT, ForeignKey('recipes.id', ondelete="CASCADE")),
+    Column('user_id', INTEGER, ForeignKey('users.id', ondelete="CASCADE")),
+    Column('recipe_id', INTEGER, ForeignKey('recipes.id', ondelete="CASCADE")),
     Column('saved_at', TIMESTAMP),
     PrimaryKeyConstraint('user_id', 'recipe_id')
 )
